refactor: Reorganize the code for reducers
the factory and resolve files for each reducer are in the same directories

#76
