+++ packages/data-point/lib/entity-types/resolve-entity/resolve-entity.js
@@ -3,7 +3,7 @@
 const _ = require('lodash')
 const Promise = require('bluebird')
 
-const middleware = require('../../resolve/middleware')
+const middleware = require('../../middleware')
 
 const utils = require('../../utils')
