+++ packages/data-point/lib/middleware/index.js
@@ -1,5 +1,6 @@
 /* eslint global-require: 0 */
 
 module.exports = {
-  create: require('./factory').create
+  create: require('./factory').create,
+  resolve: require('./resolve').resolve
 }
