+++ packages/data-point/lib/reducer-entity/resolve.js
