+++ packages/data-point/lib/reducer-function/index.js
@@ -1,9 +1,11 @@
 /* eslint global-require: 0 */
 
 const reducer = require('./factory')
+const resolve = require('./resolve').resolve
 
 module.exports = {
   create: reducer.create,
   type: reducer.type,
-  isType: reducer.isFunction
+  isType: reducer.isFunction,
+  resolve: resolve
 }
