+++ packages/data-point/lib/reducer-function/resolve.js
