+++ packages/data-point/lib/reducer-path/index.js
@@ -1,9 +1,11 @@
 /* eslint global-require: 0 */
 
 const reducer = require('./factory')
+const resolve = require('./resolve').resolve
 
 module.exports = {
   create: reducer.create,
   type: reducer.type,
-  isType: reducer.isPath
+  isType: reducer.isPath,
+  resolve: resolve
 }
