+++ packages/data-point/lib/reducer-path/resolve.js
